refactor(enspire):  __post_init__
